Module project_euler::solutions

A collection of functions for solving Project Euler questions. Based upon the parameters for the Hacker Rank implementations Currently missing: #4, #7, #13

Structs

CollatzCache

DivisorCache
Stores the number of factors a number has in a HashMap.

FibThree
Holds and implements even Fibonacci numbers

PrimeCache
Store and generate a vector of primes.

Functions

compute_multiple
Computes the smallest multiple of all the integers from 1 to n, n \in Integers.

count_even
Sums the even Fibonacci numbers up to an integer, n.

grid_prod
Move across a grid and computes products of k sequential numbers in each direction.

largest_prime_factor
Solves for the larget prime factor of an integer, n.

largest_product
Find the largest product of k sequential numbers in a sequence of n digits.

longest_collatz

pythagorean_triplets
Given N, finds the maximum product, a*b*c, of pythagorean triples {(a,b,c)| a2 + b2 = c2 } such that a+b+c = N.

squared_sum_diff
Finds the absolute difference between the sum of squares, 12 + 22 + ... + n2, and the square of the sum, (1 + 2 + ... + n)2.

sum_multiples
Sums up all the multiples of two integers [a,b] up to a certain integer n.

sum_primes
Sums up primes up to an integer n.

triangle_divisors
Finds the smallest triangular number, T_n, such that the number of factors of T_n is greater than a specified integer N.