Struct project_euler::solutions::DivisorCache

pub struct DivisorCache {
    // some fields omitted
}

Stores the number of factors a number has in a HashMap.

Methods

impl DivisorCache

fn new() -> DivisorCache

fn get(&mut self, key: i64) -> i64

Fetches the number of factors an integer has. It is calculated and stored if not currently stored.

Examples

use project_euler::solutions::DivisorCache;

let mut cache = DivisorCache::new();
assert_eq!(cache.get(7), 2); //The factors of 7 are 1,7.
assert_eq!(cache.get(28), 6); //The factors of 28 are 1,2,4,7,14,28.